+++ app/src/main/java/org/pathfinderfr/app/MainActivity.java
@@ -422,18 +422,9 @@ public class MainActivity extends AppCompatActivity
             totalCount = newEntities.size();
             factoryId = FavoriteFactory.FACTORY_ID;
         } else if (id == R.id.nav_sheet) {
-            // if character "pined", automatically select it
-            long characterId = PreferenceManager.getDefaultSharedPreferences(getBaseContext()).getLong(CharacterSheetActivity.PREF_SELECTED_CHARACTER_ID, 0L);
-            if(characterId>0 && !CharacterFactory.FACTORY_ID.equals(factoryId)) {
-                Intent intent = new Intent(this, CharacterSheetActivity.class);
-                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                intent.putExtra(CharacterSheetActivity.SELECTED_CHARACTER_ID, characterId);
-                getBaseContext().startActivity(intent);
-            } else {
-                newEntities = dbhelper.getAllEntities(CharacterFactory.getInstance());
-                totalCount = newEntities.size();
-                factoryId = CharacterFactory.FACTORY_ID;
-            }
+            newEntities = dbhelper.getAllEntities(CharacterFactory.getInstance());
+            totalCount = newEntities.size();
+            factoryId = CharacterFactory.FACTORY_ID;
         } else if (id == R.id.nav_skills) {
             newEntities = dbhelper.getAllEntities(SkillFactory.getInstance());
             totalCount = newEntities.size();